+++ Sources/ScreenShieldKit/Logger.swift
@@ -1,0 +1,10 @@
+//
+//  Logger.swift
+//  ScreenShieldKit
+//
+//  Created by Kyle on 2025/02/17.
+//
+
+import os.log
+
+let screenShieldLogger = OSLog(subsystem: "top.kyleye.screenshieldkit", category: "ScreenShieldKit")
